+++ src/hgraph/nodes/_logical.py
@@ -1,9 +1,1 @@
-from hgraph import compute_node, TS
 
-
-__all__ = ("not_",)
-
-
-@compute_node
-def not_(ts: TS[bool]) -> TS[bool]:
-    return not ts.value