+++ src/hgraph/nodes/_tss_operators.py
@@ -1,4 +1,5 @@
-from hgraph import compute_node, contains_, REF, TSS, SCALAR, TS, STATE, PythonTimeSeriesReference
+from hgraph import compute_node, contains_, REF, TSS, SCALAR, TS, STATE, PythonTimeSeriesReference, not_, graph
+from hgraph.nodes._set_operators import is_empty
 
 
 @compute_node(overloads=contains_)
@@ -7,11 +8,11 @@ def tss_contains(ts: REF[TSS[SCALAR]], item: TS[SCALAR], _state: STATE = None) \
     """Perform a time-series contains check of an item in the given time-series set"""
     # If the tss is set then we should de-register the old contains.
     if _state.tss is not None:
-        _state.tss.release_contains_ref(_state.item, _state.requester)
+        _state.tss.release_contains_output(_state.item, _state.requester)
     _state.tss = ts.value.output
     _state.item = item.value
     return PythonTimeSeriesReference(
-        None if _state.tss is None else _state.tss.get_contains_ref(_state.item, _state.requester))
+        None if _state.tss is None else _state.tss.get_contains_output(_state.item, _state.requester))
 
 
 @tss_contains.start
@@ -19,3 +20,15 @@ def _tss_contains_start(_state: STATE):
     _state.requester = object()
     _state.tss = None
     _state.item = None
+
+
+@compute_node(overloads=is_empty)
+def tss_is_empty(ts: REF[TSS[SCALAR]]) -> REF[TS[bool]]:
+    """A time-series ticking with the empty state of the TSS supplied is modified"""
+    # NOTE: Since the TSS output is currently a fixed output we don't need to track state.
+    return PythonTimeSeriesReference(ts.value.output.is_empty_output() if ts.value.valid else None)
+
+
+@graph(overloads=not_)
+def tss_not_(ts: TSS[SCALAR]) -> TS[bool]:
+    return tss_is_empty(ts)